requirements.txt
fastapi
uvicorn[standard]
python-logging-loki
python-json-logger
pydantic
python-multipart  # Required for Form data 